Add darwin memory collector
