+++ pkg/gohai/memory/memory_linux.go
@@ -7,19 +7,6 @@ import (
 	"regexp"
 )
 
-type Memory struct{}
-
-const name = "memory"
-
-func (self *Memory) Name() string {
-	return name
-}
-
-func (self *Memory) Collect() (result interface{}, err error) {
-	result, err = getMemoryInfo()
-	return
-}
-
 var memMap = map[string]string{
 	"MemTotal":  "total",
 	"SwapTotal": "swap_total",
